src/py/requirements.txt
facenet-pytorch==2.5.2
opencv-python==4.5.5.64
Pillow==9.1.1
numpy==1.22.4
argparse==1.4.0